api/requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
gitdb==4.0.9
GitPython==3.1.26
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
smmap==5.0.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1